[http-client] Move http clients to separate modules
