+++ http-client/http-client-core/Module.md
@@ -1,0 +1,3 @@
+# Module http-client-core
+
+Core HTTP client interface for Koog framework. Provides the `KoogHttpClient` interface.
