+++ http-client/http-client-core/build.gradle.kts
@@ -1,0 +1,24 @@
+import ai.koog.gradle.publish.maven.Publishing.publishToMaven
+
+plugins {
+    id("ai.kotlin.multiplatform")
+    alias(libs.plugins.kotlin.serialization)
+}
+
+group = rootProject.group
+version = rootProject.version
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(libs.kotlinx.coroutines.core)
+                api(libs.jetbrains.annotations)
+            }
+        }
+    }
+
+    explicitApi()
+}
+
+publishToMaven()
