+++ http-client/http-client-java/Module.md
@@ -1,0 +1,3 @@
+# Module http-client-java
+
+Java 11+ HttpClient-based implementation of `KoogHttpClient` for JVM platforms.
