+++ http-client/http-client-ktor/Module.md
@@ -1,0 +1,3 @@
+# Module http-client-ktor
+
+Ktor-based implementation of `KoogHttpClient` for multiplatform HTTP operations.
