+++ http-client/http-client-okhttp/Module.md
@@ -1,0 +1,3 @@
+# Module http-client-okhttp
+
+OkHttp-based implementation of `KoogHttpClient` for JVM platforms.
