+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-openai-client-base/build.gradle.kts
@@ -14,6 +14,7 @@ kotlin {
                 api(project(":utils"))
                 api(project(":prompt:prompt-executor:prompt-executor-clients"))
                 api(project(":prompt:prompt-structure"))
+                api(project(":http-client:http-client-ktor"))
                 api(libs.ktor.client.content.negotiation)
                 api(libs.ktor.serialization.kotlinx.json)
                 implementation(libs.oshai.kotlin.logging)
